pyproject.toml
[project]
name = "kalshi-edge"
version = "0.1.0"
description = "Probabilistic forecasting and edge detection for Kalshi prediction markets."
requires-python = ">=3.10,<3.13"
authors = [{ name = "Andrew" }]
dependencies = [
    # HTTP / caching / retries
    "requests>=2.31",
    "requests-cache>=1.2",
    "tenacity>=8.2",
    # Config
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    # Numerical core
    "numpy>=1.26,<2.1",
    "scipy>=1.12",
    "pandas>=2.2",
    "polars>=0.20",
    # Stats / time series
    "statsmodels>=0.14",
    "arch>=6.3",
    # Bayesian inference (CPU-only JAX by default)
    "numpyro>=0.14",
    "jax>=0.4.25",
    "jaxlib>=0.4.25",
    # Cryptography (Kalshi RSA request signing)
    "cryptography>=42.0",
    # Scheduler
    "apscheduler>=3.10",
    # Dashboard
    "streamlit>=1.32",
    "plotly>=5.19",
    "matplotlib>=3.8",
    # Email templating
    "jinja2>=3.1",
    # CLI / logging
    "rich>=13.7",
    "click>=8.1",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-xdist>=3.5",
    "hypothesis>=6.98",
    "responses>=0.25",
    "ruff>=0.3",
    "mypy>=1.9",
    "types-PyYAML",
    "types-requests",
]

[project.scripts]
kalshi-edge = "kalshi_edge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kalshi_edge"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
filterwarnings = [
    "ignore::DeprecationWarning:jax.*",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "N", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
disallow_untyped_defs = true
warn_unused_ignores = true
ignore_missing_imports = true
